pyproject.toml
[project]
name = "sdwis-automation"
version = "0.1.0"
description = "Mississippi SDWIS web automation for water system data extraction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "playwright>=1.55.0",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
]
